getJobById: Get a single job by Id. If job not found, returns a dummy job (not ideal ... will fix later).

+++ src/main/java/com/example/firstjobapp/job/Job.java
@@ -23,7 +23,7 @@ public class Job {
 
     // ---------- Getters / Setters ----------------
     public Long getId() {
-        return id;
+        return (long) id;
     }
 
     public void setId(Long id) {
